Added locationPermission rule. Added null checks in clearPreferences and clearDatabase methods.

+++ AndroidDemo/app/src/androidTest/java/com/mytaxi/android_demo/tests/Tests.kt
@@ -31,34 +31,39 @@ open class Tests {
     fun clearPreferences() {
         LogT.d("clear preferences")
         val root = appContext.filesDir.parentFile
-        val sharedPreferencesFileNames = File(root, "shared_prefs").list()
-        for (fileName in sharedPreferencesFileNames) {
-            val sharedPreferencesName = fileName.replace(".xml", "")
-            //delete doesn't seem to work
-            appContext.getSharedPreferences(sharedPreferencesName, Context.MODE_PRIVATE).edit().clear().commit()
+        val sharedPreferencesFileNames: Array<String>? = File(root, "shared_prefs").list()
+        if (sharedPreferencesFileNames != null) {
+            for (fileName in sharedPreferencesFileNames) {
+                val sharedPreferencesName = fileName.replace(".xml", "")
+                //delete doesn't seem to work
+                appContext.getSharedPreferences(sharedPreferencesName, Context.MODE_PRIVATE).edit().clear().commit()
+            }
         }
     }
 
     fun clearDatabase() {
         LogT.d("clear database")
-        for (database in appContext.databaseList()) {
-            // when transaction rollback files exists they are always locked so we can't delete them
-            if (database.contains(".db-journal")) {
-                appContext.deleteDatabase(database)
-                continue
-            }
+        val databaseList: Array<String>? = appContext.databaseList()
+        if (databaseList != null) {
+            for (database in databaseList) {
+                // when transaction rollback files exists they are always locked so we can't delete them
+                if (database.contains(".db-journal")) {
+                    appContext.deleteDatabase(database)
+                    continue
+                }
 
-            // when using transaction write ahead logging then this db files are listed but often they don't exist
-            if (database.contains(".db-wal") || database.contains(".db-shm")) {
-                appContext.deleteDatabase(database)
-                continue
-            }
+                // when using transaction write ahead logging then this db files are listed but often they don't exist
+                if (database.contains(".db-wal") || database.contains(".db-shm")) {
+                    appContext.deleteDatabase(database)
+                    continue
+                }
 
-            LogT.d("deleting $database")
+                LogT.d("deleting $database")
 
-            val databasePath = appContext.getDatabasePath(database)
-            if (databasePath.exists()) {
-                appContext.deleteDatabase(database)
+                val databasePath = appContext.getDatabasePath(database)
+                if (databasePath.exists()) {
+                    appContext.deleteDatabase(database)
+                }
             }
         }
     }
